requirements.txt
absl-py==0.8.1
actionlib==1.11.13
anaconda==0.0.1.1
angles==1.9.11
appdirs==1.4.3
apturl==0.5.2
astor==0.8.0
astroid==2.3.1
baidu-api==0.0.2
beautifulsoup4==4.4.1
blinker==1.3
bondpy==1.8.3
Brlapi==0.6.4
cachetools==4.0.0
camera-calibration-parsers==1.11.13
capabilities==0.2.0
catkin==0.7.18
catkin-pkg==0.4.13
catkin-pkg-modules==0.4.16
catkin-tools==0.4.5
certifi==2019.9.11
-e git+https://github.com/bitcraze/crazyflie-clients-python.git@02181db7118f9bb2cefbad624499d90efffdfcf6#egg=cfclient
-e git+https://github.com/bitcraze/crazyflie-lib-python.git@3c24e993bfa8eb8b3710bd1a3577e647aab6d7e3#egg=cflib
chardet==3.0.4
checkbox-support==0.22
Click==7.0
colorama==0.4.1
command-not-found==0.3
controller-manager==0.13.3
controller-manager-msgs==0.13.3
create-dashboard==2.3.1
create-driver==2.3.1
create-node==2.3.1
cryptography==1.2.3
cycler==0.10.0
Cython==0.29.13
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
diagnostic-updater==1.9.3
dlib==19.17.0
docutils==0.15.2
dynamic-reconfigure==1.5.50
empy==3.3.4
face-recognition==1.2.3
face-recognition-models==0.3.0
feedparser==5.1.3
fire==0.2.1
gast==0.3.2
gazebo-ros==2.5.19
gencpp==0.6.0
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.11
gennodejs==2.0.1
genpy==0.6.7
google-api-core==1.16.0
google-auth==1.13.1
google-cloud-language==1.3.0
google-cloud-speech==1.3.2
google-pasta==0.1.7
googleapis-common-protos==1.51.0
grpcio==1.24.1
guacamole==0.9.2
h5py==2.10.0
html5lib==0.999
httplib2==0.9.1
idna==2.8
imageio==2.8.0
isort==4.3.21
Jinja2==2.8
kazam==1.4.5
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
kobuki-dashboard==0.5.7
kobuki-testsuite==0.7.6
language-selector==0.1
laser-geometry==1.6.4
lazy-object-proxy==1.4.2
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
Markdown==3.1.1
MarkupSafe==0.23
matplotlib==3.0.3
mccabe==0.6.1
message-filters==1.12.14
netifaces==0.10.9
networkx==2.4
numpy==1.17.3
oauthlib==1.0.3
onboard==1.2.0
opencv-python==4.1.1.26
openni==2.3.0
osrf-pycommon==0.1.9
padme==1.1.1
pandas==0.23.4
pathspec==0.6.0
pexpect==4.0.1
Pillow==6.1.0
pipdate==0.3.1
plainbox==0.25
playsound==1.2.2
pluginlib==1.11.3
protobuf==3.10.0
psutil==5.6.3
ptyprocess==0.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
PyBrain==0.3
pycocotools==2.0.0
pycups==1.9.73
pycurl==7.43.0
Pygments==2.1
pygobject==3.20.0
PyJWT==1.3.0
pylibfreenect2==0.1.3
pylint==2.4.2
pyparsing==2.4.2
pyqtgraph==0.10.0
python-apt==1.1.0b1+ubuntu0.16.4.8
python-dateutil==2.8.0
python-debian==0.1.27
python-qt-binding==0.3.4
python-systemd==231
pytz==2019.2
pyusb==1.0.0b2
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==5.1.2
pyzmq==18.1.0
qt-gui==0.3.11
qt-gui-cpp==0.3.11
qt-gui-py-common==0.3.11
qtm==2.0.3
Quamash==0.6.1
redis==3.3.11
reportlab==3.3.0
requests==2.22.0
rocon-app-manager==0.9.1
rocon-app-utilities==0.9.1
rocon-apps==0.9.1
rocon-console==0.3.2
rocon-ebnf==0.3.2
rocon-gateway==0.8.1
rocon-gateway-utils==0.8.1
rocon-hub==0.8.1
rocon-hub-client==0.8.1
rocon-interactions==0.3.2
rocon-master-info==0.3.2
rocon-python-comms==0.3.2
rocon-python-redis==0.3.2
rocon-python-utils==0.3.2
rocon-python-wifi==0.3.2
rocon-semantic-version==0.3.2
rocon-uri==0.3.2
roman==2.0.0
rosbag==1.12.14
rosboost-cfg==1.14.6
rosclean==1.14.6
roscreate==1.14.6
rosdep==0.16.1
rosdistro==0.7.4
rosgraph==1.12.14
rosinstall==0.7.8
rosinstall-generator==0.1.17
roslaunch==1.12.14
roslib==1.14.6
roslz4==1.12.14
rosmake==1.14.6
rosmaster==1.12.14
rosmsg==1.12.14
rosnode==1.12.14
rosparam==1.12.14
rospkg==1.1.10
rospy==1.12.14
rospy-message-converter==0.5.0
rosservice==1.12.14
rostest==1.12.14
rostopic==1.12.14
rosunit==1.14.6
roswtf==1.12.14
rqt-bag==0.4.12
rqt-console==0.4.8
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-logger-level==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.8
rqt-py-common==0.5.0
rqt-reconfigure==0.5.1
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.8
rqt-wrapper==0.1.4
rsa==4.0
scikit-image==0.15.0
scipy==1.3.1
screen-resolution-extra==0.0.0
sensor-msgs==1.12.7
sessioninstaller==0.0.0
six==1.12.0
slidingwindow==0.0.13
smclib==1.8.3
sound-play==0.3.3
SpeechRecognition==3.8.1
ssh-import-id==5.5
statistics==1.0.3.5
structure==0.1.dev16
system-service==0.3
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorpack==0.9.8
termcolor==1.1.0
test-py==0.3
tf==1.11.9
tf-conversions==1.11.9
tf-pose==0.1.1
tf2-py==0.5.20
tf2-ros==0.5.20
tinynetrc==1.3.0
topic-tools==1.12.14
tqdm==4.36.1
trollius==2.1.post2
typed-ast==1.4.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unique-id==1.0.5
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.25.6
usb-creator==0.3.0
vcstools==0.1.42
virtualenv==15.0.1
Werkzeug==0.16.0
world-canvas-server==0.2.0
world-canvas-utils==0.2.0
wrapt==1.11.2
wstool==0.1.17
xacro==1.11.3
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
